Index into vectors using integers.
This commit corrects an error that has been present since the
introduction of function wellsToSrc() in commit a50bb8f but was never
detected.  Although the conversion int->double->std::size_t is likely to
be loss-less, it is better to not introduce any more steps than are
actually needed.

+++ opm/core/utility/miscUtilities.cpp
@@ -474,7 +474,7 @@ namespace Opm
             if (wells.type[w] == INJECTOR) {
                 flow *= wells.comp_frac[np*w + 0]; // Obtaining water rate for inflow source.
             }
-            const double cell = wells.well_cells[wells.well_connpos[w]];
+            const int cell = wells.well_cells[wells.well_connpos[w]];
             src[cell] = flow;
         }
     }
